chore: moved directory location of Fabric component
