+++ src/index.ts
@@ -1,5 +1,5 @@
 import Video from './Video';
-export {VideoDecoderProperties} from './VideoNativeComponent';
+export {VideoDecoderProperties} from './specs/VideoNativeComponent';
 export * from './types';
 export type {VideoRef} from './Video';
 export default Video;
